requirements.txt
numpy
pandas
matplotlib
yfinance
djangorestframework
xlwt
psycopg2